chore: Option type
