+++ packages/fuselage/src/components/Options/Option/Option.stories.mdx
@@ -2,8 +2,7 @@ import { action } from '@storybook/addon-actions';
 import { Meta, Canvas, ArgsTable, Story } from '@storybook/addon-docs';
 import LinkTo from '@storybook/addon-links/react';
 
-import { Option } from '.';
-import { Box, Avatar, Menu, StatusBullet } from '../..';
+import { Box, Avatar, Menu, StatusBullet, Option } from '../..';
 import { exampleAvatar, menuOptions } from '../../../../.storybook/helpers.js';
 
 <Meta title='Misc/Options/Option' parameters={{ jest: ['Option/spec'] }} />
