+++ packages/fuselage/src/components/index.ts
@@ -25,7 +25,7 @@ export * from './Modal';
 export * from './MultiSelect';
 export * from './NumberInput';
 export * from './Options';
-export * from './Options/Option';
+export { default as Option } from './Options/Option';
 export * from './Pagination';
 export { default as PasswordInput } from './PasswordInput';
 export * from './StatusBullet';
